Split off scenery entries
BannerSceneryEntry, WallSceneryEntry footpathItemEntry
