+++ src/openrct2/entity/Guest.cpp
@@ -29,7 +29,9 @@
 #include "../management/Marketing.h"
 #include "../management/NewsItem.h"
 #include "../network/network.h"
+#include "../object/FootpathItemEntry.h"
 #include "../object/MusicObject.h"
+#include "../object/WallSceneryEntry.h"
 #include "../peep/GuestPathfinding.h"
 #include "../peep/RideUseSystem.h"
 #include "../rct2/RCT2.h"
