+++ src/openrct2/interface/Viewport.cpp
@@ -23,6 +23,7 @@
 #include "../entity/PatrolArea.h"
 #include "../entity/Staff.h"
 #include "../object/SmallSceneryEntry.h"
+#include "../object/WallSceneryEntry.h"
 #include "../paint/Paint.h"
 #include "../profiling/Profiling.h"
 #include "../ride/Ride.h"
