+++ src/openrct2/paint/tile_element/Paint.Wall.cpp
@@ -17,6 +17,7 @@
 #include "../../interface/Viewport.h"
 #include "../../localisation/Formatting.h"
 #include "../../localisation/Localisation.h"
+#include "../../object/WallSceneryEntry.h"
 #include "../../profiling/Profiling.h"
 #include "../../ride/Track.h"
 #include "../../ride/TrackDesign.h"
