+++ src/openrct2/world/ConstructionClearance.cpp
@@ -12,6 +12,7 @@
 #include "../Game.h"
 #include "../localisation/Formatter.h"
 #include "../object/SmallSceneryEntry.h"
+#include "../object/WallSceneryEntry.h"
 #include "../openrct2/Cheats.h"
 #include "../ride/Ride.h"
 #include "../ride/RideData.h"
