+++ src/openrct2/world/Scenery.cpp
@@ -23,9 +23,11 @@
 #include "../entity/Fountain.h"
 #include "../localisation/Localisation.h"
 #include "../network/network.h"
+#include "../object/FootpathItemEntry.h"
 #include "../object/ObjectList.h"
 #include "../object/ObjectManager.h"
 #include "../object/SmallSceneryEntry.h"
+#include "../object/WallSceneryEntry.h"
 #include "../scenario/Scenario.h"
 #include "Climate.h"
 #include "Footpath.h"
