+++ src/openrct2/world/TileElement.cpp
@@ -12,6 +12,7 @@
 #include "../core/Guard.hpp"
 #include "../interface/Window.h"
 #include "../localisation/Localisation.h"
+#include "../object/WallSceneryEntry.h"
 #include "../ride/Track.h"
 #include "Banner.h"
 #include "LargeScenery.h"
